Refactor existing upload ballot logic into a sub-component
