+++ src/electionguard_gui/web/components/admin/upload-ballots-wizard-component.js
@@ -1,0 +1,3 @@
+export default {
+  template: /*html*/ `<h1>Upload Wizard</h1>`,
+};
